W-12967336: In core move some claasses in processor and routing to a different/new package
