+++ core/src/main/java/org/mule/runtime/core/api/transaction/TransactionCoordination.java
@@ -8,7 +8,7 @@ package org.mule.runtime.core.api.transaction;
 
 import org.mule.runtime.api.tx.TransactionException;
 import org.mule.runtime.core.api.config.i18n.CoreMessages;
-import org.mule.runtime.core.internal.processor.DelegateTransaction;
+import org.mule.runtime.core.internal.transaction.DelegateTransaction;
 import org.mule.runtime.core.privileged.transaction.xa.IllegalTransactionStateException;
 
 import java.util.ArrayDeque;
