+++ core/src/main/java/org/mule/runtime/core/internal/exception/ErrorTypeLocatorFactory.java
@@ -41,7 +41,7 @@ import org.mule.runtime.core.api.transformer.MessageTransformerException;
 import org.mule.runtime.core.api.transformer.TransformerException;
 import org.mule.runtime.core.internal.construct.FlowBackPressureException;
 import org.mule.runtime.core.internal.event.EventContextDeepNestingException;
-import org.mule.runtime.core.internal.routing.DuplicateMessageException;
+import org.mule.runtime.core.internal.routing.split.DuplicateMessageException;
 import org.mule.runtime.core.internal.routing.ValidationException;
 import org.mule.runtime.core.privileged.exception.ErrorTypeLocator;
 import org.mule.runtime.core.privileged.exception.MessageRedeliveredException;
