+++ core/src/main/java/org/mule/runtime/core/internal/routing/outbound/CollectionMessageSequence.java
@@ -8,8 +8,8 @@ package org.mule.runtime.core.internal.routing.outbound;
 
 import static java.util.Objects.requireNonNull;
 
-import org.mule.runtime.core.internal.routing.AbstractMessageSequence;
-import org.mule.runtime.core.internal.routing.MessageSequence;
+import org.mule.runtime.core.internal.routing.split.AbstractMessageSequence;
+import org.mule.runtime.core.internal.routing.split.MessageSequence;
 
 import java.util.Collection;
 import java.util.Iterator;
