+++ core/src/main/java/org/mule/runtime/core/internal/routing/outbound/NodeListMessageSequence.java
@@ -6,7 +6,7 @@
  */
 package org.mule.runtime.core.internal.routing.outbound;
 
-import org.mule.runtime.core.internal.routing.AbstractMessageSequence;
+import org.mule.runtime.core.internal.routing.split.AbstractMessageSequence;
 
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
